# Switch to persistent adaptive radix trees