Switch StateDB to use custom implementation of Persistent Adaptive Radix Trees (https://db.in.tum.de/~leis/papers/ART.pdf).

The implementation isn't a completely fateful implementation of the paper. The Node16 SIMD tricks and Node48 key array are skipped. This made the implementation simpler, but did leave some performance improvements on the table.

The only functional change to StateDB is to drop the `Last` and `LastWatch` methods since that requires an reverse iterator implementation and since we didn't use them we can just drop them for now.

Benchmark before:

```
goos: linux                                                                                                                                                                                     
goarch: amd64                                                                                                                                                                                   
pkg: github.com/cilium/statedb                                                                                                                                                                  
cpu: Intel(R) Core(TM) i5-10210U CPU @ 1.60GHz                                                                                                                                                  
BenchmarkDB_WriteTxn_1-8                                          307952              4012 ns/op            249227 objects/sec      3080 B/op         53 allocs/op                              
BenchmarkDB_WriteTxn_10-8                                         504811              2373 ns/op            421360 objects/sec      1501 B/op         23 allocs/op                              
BenchmarkDB_WriteTxn_100-8                                        522646              2229 ns/op            448544 objects/sec      1399 B/op         19 allocs/op                              
BenchmarkDB_WriteTxn_1000-8                                       369973              2954 ns/op            338547 objects/sec      1492 B/op         18 allocs/op                              
BenchmarkDB_WriteTxn_10000-8                                      995700              2792 ns/op            358210 objects/sec      1355 B/op         21 allocs/op                              
BenchmarkDB_WriteTxn_100_SecondaryIndex-8                         293709              3843 ns/op            260247 objects/sec      2208 B/op         33 allocs/op                              
BenchmarkDB_RandomInsert-8                                          1032           1192822 ns/op            838348 objects/sec    796015 B/op      12214 allocs/op                              
BenchmarkDB_RandomReplace-8                                          243           4906222 ns/op            203823 objects/sec   2382866 B/op      33380 allocs/op                              
BenchmarkDB_SequentialInsert-8                                       430           2770646 ns/op            360927 objects/sec   1492345 B/op      18256 allocs/op                              
BenchmarkDB_Baseline_SingleRadix_Insert-8                           3109            371203 ns/op           2693945 objects/sec    353662 B/op       5103 allocs/op                              
BenchmarkDB_Baseline_SingleRadix_TrackMutate_Insert-8               3105            373685 ns/op           2676056 objects/sec    353951 B/op       5106 allocs/op                              
BenchmarkDB_Baseline_SingleRadix_Lookup-8                          13185             90097 ns/op          11099219 objects/sec         0 B/op          0 allocs/op                              
BenchmarkDB_Baseline_Part_RootOnlyWatch_Insert-8                    5473            204824 ns/op           4882248 objects/sec    146969 B/op       3065 allocs/op                              
BenchmarkDB_Baseline_Part_Insert-8                                  4282            257142 ns/op           3888905 objects/sec    246580 B/op       4094 allocs/op                              
BenchmarkDB_Baseline_Part_Lookup-8                                 36069             33324 ns/op          30008204 objects/sec         0 B/op          0 allocs/op                              
BenchmarkDB_Baseline_Hashmap_Insert-8                              15986             74292 ns/op          13460333 objects/sec     86561 B/op         64 allocs/op                              
BenchmarkDB_Baseline_Hashmap_Lookup-8                              84038             14228 ns/op          70282379 objects/sec         0 B/op          0 allocs/op                              
BenchmarkDB_DeleteTracker_Baseline-8                                 385           3065639 ns/op            326197 objects/sec   1758394 B/op      23413 allocs/op                              
BenchmarkDB_DeleteTracker-8                                          192           6763615 ns/op            147850 objects/sec   3387030 B/op      40807 allocs/op                              
BenchmarkDB_RandomLookup-8                                          7114            153828 ns/op           6500774 objects/sec       144 B/op          1 allocs/op                              
BenchmarkDB_SequentialLookup-8                                      8168            142537 ns/op           7015728 objects/sec      8000 B/op       1000 allocs/op                              
BenchmarkDB_FullIteration_All-8                                   100930             11497 ns/op          86979024 objects/sec       280 B/op          6 allocs/op                              
BenchmarkDB_FullIteration_Get-8                                    81974             14355 ns/op          69661408 objects/sec       320 B/op          7 allocs/op
BenchmarkDB_PropagationDelay-8                                    243086              4838 ns/op                43.00 50th_µs           50.00 90th_µs          212.0 99th_µs        3083 B/op        50 allocs/op
```

Benchmark after:

```
BenchmarkDB_WriteTxn_1-8                                  380911              3093 ns/op            323322 objects/sec      2440 B/op         38 allocs/op
BenchmarkDB_WriteTxn_10-8                                 901350              1336 ns/op            748741 objects/sec       699 B/op         12 allocs/op
BenchmarkDB_WriteTxn_100-8                               1000000              1075 ns/op            930477 objects/sec       581 B/op          9 allocs/op
BenchmarkDB_WriteTxn_1000-8                              1000000              1262 ns/op            792304 objects/sec       578 B/op          9 allocs/op
BenchmarkDB_WriteTxn_10000-8                             1000000              1262 ns/op            792171 objects/sec       548 B/op          9 allocs/op
BenchmarkDB_WriteTxn_100_SecondaryIndex-8                 705823              1752 ns/op            570791 objects/sec       934 B/op         18 allocs/op
BenchmarkDB_RandomInsert-8                                  1563            743594 ns/op           1344819 objects/sec    481777 B/op       9162 allocs/op
BenchmarkDB_RandomReplace-8                                  530           2266878 ns/op            441135 objects/sec    932101 B/op      19170 allocs/op
BenchmarkDB_SequentialInsert-8                               970           1264837 ns/op            790616 objects/sec    578330 B/op       9168 allocs/op
BenchmarkDB_Baseline_Part_RootOnlyWatch_Insert-8            5121            220673 ns/op           4531599 objects/sec    146970 B/op       3065 allocs/op
BenchmarkDB_Baseline_Part_Insert-8                          4135            281535 ns/op           3551957 objects/sec    246583 B/op       4094 allocs/op
BenchmarkDB_Baseline_Part_Lookup-8                         35305             34346 ns/op          29115518 objects/sec         0 B/op          0 allocs/op
BenchmarkDB_Baseline_Hashmap_Insert-8                      14404             83092 ns/op          12034837 objects/sec     86549 B/op         64 allocs/op
BenchmarkDB_Baseline_Hashmap_Lookup-8                      77744             15222 ns/op          65694567 objects/sec         0 B/op          0 allocs/op
BenchmarkDB_DeleteTracker_Baseline-8                         945           1267367 ns/op            789038 objects/sec    648963 B/op      13062 allocs/op
BenchmarkDB_DeleteTracker-8                                  501           2361545 ns/op            423452 objects/sec   1176999 B/op      20188 allocs/op
BenchmarkDB_RandomLookup-8                                 13999             82787 ns/op          12079196 objects/sec       144 B/op          1 allocs/op
BenchmarkDB_SequentialLookup-8                             14426             77552 ns/op          12894616 objects/sec      8000 B/op       1000 allocs/op
BenchmarkDB_FullIteration_All-8                            86900             12131 ns/op          82435475 objects/sec       264 B/op          6 allocs/op
BenchmarkDB_FullIteration_Get-8                            75100             14046 ns/op          71195485 objects/sec       304 B/op          7 allocs/op
BenchmarkDB_PropagationDelay-8                            398498              3135 ns/op                27.00 50th_µs           32.00 90th_µs          170.0 99th_µs        1479 B/op         28 allocs/op
```

Highlights:
- WriteTxn_1 went from 250k/s to 323k/s
- WriteTxn_100 went from 448k/s to 930k/s
- RandomLookup from 700k/s to 1208k/s
- DeleteTracker up from 150k/s to 423k/s
- Propagation delay down from 43uS to 27uS (50th)
- Reconciler throughput up from ~200k/s to 350k/s, memory usage down 50%