@joamaki 2024-04-09
Benchmark results vs immutable-radix:

```
BenchmarkDB_Baseline_Part_Insert-8                      	   5487	   204858 ns/op	  4881440 objects/sec
BenchmarkDB_Baseline_SingleRadix_Insert-8               	   3091	   375413 ns/op	  2663733 objects/sec
BenchmarkDB_Baseline_SingleRadix_TrackMutate_Insert-8   	   2935	   377532 ns/op	  2648787 objects/sec
BenchmarkDB_Baseline_Part_TrackMutate_Insert-8          	   4450	   255397 ns/op	  3915473 objects/sec
BenchmarkDB_Baseline_SingleRadix_Lookup-8               	  12249	    91621 ns/op	 10914531 objects/sec
BenchmarkDB_Baseline_Part_Lookup-8                      	  12066	    97336 ns/op	 10273722 objects/sec
BenchmarkDB_Baseline_Hashmap_Insert-8                   	  15342	    77111 ns/op	 12968333 objects/sec
BenchmarkDB_Baseline_Hashmap_Lookup-8                   	  83860	    14259 ns/op	 70132148 objects/sec
```

"Part_Insert" is without the watch channel management, and "Part_TrackMutate_Insert" is with. "SingleRadix_TrackMutate" still creates the watch channels but doesn't close them (so it's not apples-to-apples).