@joamaki 2024-04-11
The impact of this is really visible in the reconciler benchmark:
```
benchmark % go run . -batchsize 100                                                                                                                                    (pr/joamaki/part)statedb
Inserting batch 1000/1000 ...
Waiting for reconciliation to finish ...

100000 objects reconciled in 0.48 seconds (batch size 100)
Throughput 206555.64 objects per second
Allocated 1423417 objects, 115971kB bytes, 117728kB bytes still in use
benchmark % stg push
Pushing patch "statedb-switch-to-part" ... done (unmodified)
Now at patch "statedb-switch-to-part"
benchmark % go run . -batchsize 100
Inserting batch 1000/1000 ...
Waiting for reconciliation to finish ...

100000 objects reconciled in 0.34 seconds (batch size 100)
Throughput 297510.35 objects per second
Allocated 983554 objects, 57910kB bytes, 60208kB bytes still in use
```

It's 50% faster and we're using almost half the amount of memory for the objects.