@joamaki 2024-04-15
> Found a few small thing. In general things seem to check out. However, the part/iterator.go and part/txn.go have some very dense code which I find difficult to reason through. That is partially due to also having to wrap my head around the ART implementation which is new to me.
> 
> There are a few edge cases in there which are not covered by the current tests, so I would like to schedule a codewalk to go through it together before my final verdict.

Thanks for the review! What edge cases are not covered? I think it'd be great to have someone else besides me write a few test cases for this and try to break it...